# Remove an unused leaser.DB call in pebble_cache.Write