--- a/enterprise/server/backends/pebble_cache/pebble_cache.go
+++ b/enterprise/server/backends/pebble_cache/pebble_cache.go
@@ -2254,11 +2254,6 @@ func (p *PebbleCache) Writer(ctx context.Context, r *rspb.ResourceName) (interfa
 			attribute.Int64("digest_size", r.GetDigest().GetSizeBytes()))
 	}
 	defer spn.End()
-	db, err := p.leaser.DB()
-	if err != nil {
-		return nil, err
-	}
-	defer db.Close()
 
 	// If data is not already compressed, return a writer that will compress it before writing
 	// Only compress data over a given size for more optimal compression ratios